Fix timestamps on events
The event was reporting some crazy numbers from before 1970 and
sometimes with a 10 year difference between events that took place
within seconds.

The code was losing information as it was forcing a 64 bit number into a
32 bit number. This fix provides the correct timestamp in
nanoseconds.

+++ src/gpio_chip.c
@@ -97,7 +97,7 @@ int chip_read_event_data(struct gpio_chip_event_handle *handle, struct gpio_chip
         return -1;
 
     data->timestamp = event_data.timestamp;
-    data->id = (int)event_data.id;
+    data->id = event_data.id;
 
     return 0;
 }
